Add event handling features to WebRTC plugin
This commit adds several event handling features to our WebRTC plugin. These updates streamline the response to client requests, such as creating a new thread, broadcasting, answering, and handling ice candidate requests. For better clarity and structure, part of the logic was moved from the 'WebSocketEndpoint' class to the 'WebRTCGStreamerPlugInEventsHandler' class. This should improve maintainability and make it easier to understand and update the code in the future.
