+++ src/main/java/africa/jopen/ripple/sockets/WebsocketEndpoint.java
@@ -131,193 +131,43 @@ public class WebsocketEndpoint implements WsListener {
 							break;
 						case "iceCandidate":
 							
-							if (Objects.isNull(client)) {
-								log.info("Client not found");
-								break;
+							if (!Objects.isNull(client)
+									&& jsonObject.has("feature")
+									&& jsonObject.getString("feature").equals("G_STREAM_BROADCAST")
+							) {
+								WebRTCGStreamerPlugInEventsHandler.iceCandidate(session, client, jsonObject, transaction, last);
 							}
-							if (!jsonObject.has("candidate")) {
-								log.info("Required candidate not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required candidate not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
-							}
-							if (!jsonObject.has("sdpMLineIndex")) {
-								log.info("Required sdpMLineIndex not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required sdpMLineIndex not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
-							}
-							
-							if (!jsonObject.has("threadRef")) {
-								log.info("Required threadRef not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
-							}
-							plugin = client.getWebRTCStreamMap().get(jsonObject.getInt("threadRef"));
-							if (Objects.isNull(plugin)) {
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef Value not found,Reference error")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								break;
-							}
-							plugin.setTransaction(transaction);
-							plugin.handleIceSdp(jsonObject.getString("candidate"), jsonObject.getInt("sdpMLineIndex"));
 							
 							break;
 						
 						
 						case "answer":
-							if (Objects.isNull(client)) {
-								log.info("Client not found");
-								break;
-							}
-							if (!jsonObject.has("answer")) {
-								log.info("Required answer not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required answer not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
+							if (!Objects.isNull(client)
+									&& jsonObject.has("feature")
+									&& jsonObject.getString("feature").equals("G_STREAM_BROADCAST")
+							) {
+								WebRTCGStreamerPlugInEventsHandler.answer(session, client, jsonObject, transaction, last);
 							}
 							
-							
-							if (!jsonObject.has("threadRef")) {
-								log.info("Required threadRef not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
-							}
-							plugin = client.getWebRTCStreamMap().get(jsonObject.getInt("threadRef"));
-							
-							if (Objects.isNull(plugin)) {
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef Value not found,Reference error")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								break;
-							}
-							plugin.setTransaction(transaction);
-							plugin.handleSdp(jsonObject.getString("answer"));
 							break;
 						case "startBroadCast":
-							if (Objects.isNull(client)) {
-								log.info("Client not found");
-								break;
+							if (!Objects.isNull(client)
+									&& jsonObject.has("feature")
+									&& jsonObject.getString("feature").equals("G_STREAM_BROADCAST")
+							) {
+								WebRTCGStreamerPlugInEventsHandler.startBroadCast(session, client, jsonObject, transaction, last);
 							}
-							if (!jsonObject.has("threadRef")) {
-								log.info("Required threadRef not found,Invalid request");
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef not found,Invalid request")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								
-								break;
-							}
-							plugin = client.getWebRTCStreamMap().get(jsonObject.getInt("threadRef"));
-							if (Objects.isNull(plugin)) {
-								session.send(
-										new JSONObject()
-												.put("success", false)
-												.put("eventType", jsonObject.getString("requestType"))
-												.put("accessAuth", "GENERAL")
-												.put("message", "Required threadRef Value not found,Reference error")
-												.put("transaction", transaction)
-												.put("clientID", jsonObject.getString("clientID"))
-												.toString()
-										, last);
-								break;
-							}
-							plugin.setTransaction(transaction);
-							plugin.startCall();
+							
 							break;
 						case "newThread":
-							if (Objects.isNull(client)) {
-								log.info("Client not found");
-								break;
+							if (!Objects.isNull(client)
+									&& jsonObject.has("feature")
+									&& jsonObject.getString("feature").equals("G_STREAM_BROADCAST")
+							) {
+								client.setDebugSession(isDebugSession);
+								WebRTCGStreamerPlugInEventsHandler.newThread(session, client, jsonObject, transaction, last);
 							}
-							client.setDebugSession(isDebugSession);
-							int thread = 0;
-							JSONObject reqstData = jsonObject.getJSONObject("data");
-							if (jsonObject.getString("feature").equals("G_STREAM_BROADCASTER")) {
-								
-								var filePath = XUtils.MAIN_CONFIG_MODEL.storagePath().media() + reqstData.getString("file");
-								
-								if (!new File(filePath).exists()) {
-									log.info("File not found: " + filePath);
-									
-									client.replyToNewThreadInvalidRequest(transaction, -1);
-									break;
-									
-								}
-								thread = client.createAccessGStreamerPlugIn(
-										new MediaFile(filePath
-												, 0));
-								
-							}
-							client.replyToNewThreadRequest(transaction, thread);
+							
 							break;
 						case "remember":
 							log.info("client: " + client);
